# alts: receive low watermark support