@arjan-bal 2025-10-03
> I don't think using a custom dialer would work since we need to update the value of socket option before every read. Maybe we should consider directly implementing this in the transport.


The socket option should be set to the maximum of the HTTP/2 frame size and the TLS/ALTS record size: `max(http/2_frame_size, tls_record_size)`. This would be similar to the gRPC C-core API, which [passes a parameter specifically for this purpose](https://github.com/grpc/grpc/blob/311ea6e4a52be9e8acfead4f7df22f9e9e99552e/src/core/lib/iomgr/endpoint.h#L59-L61). This value would need to be propagated down through the layers, from the framer to the credentials, and finally to the dialer (`framer -> credentials -> dialer`).

However, there are several challenges to implementing this in gRPC-Go:
* **`net.Conn` Interface**: The standard `net.Conn` interface, which gRPC-Go uses for socket operations, does not support passing a minimum read size. A potential workaround is to define an optional interface that our connection types can implement, which we can then check for using a type assertion.
* **HTTP/2 Framer**: gRPC-Go uses the standard library's HTTP/2 framer. Exposing the frame size information would likely require forking and modifying this package.
* **TLS Record Size**: Similarly, the standard library's TLS implementation doesn't easily expose the record size, which is another necessary input for this logic.