@kevinGC 2025-10-03
I actually started by writing my own `net.Conn`, but it didn't make sense: you have to call a bunch of special methods on it that make it confusing to use. I agree it'd be cleaner if it were possible, but right now that's not the case.

I'm not sure I understand the need for HTTP2 frame and TLS record size. HTTP2 is the layer above this; we care about the ALTS frame size. And IIUC ALTS requires the whole record in order to decrypt, so I'm not sure why we need the TLS record size. But please correct me if I'm wrong here; I'm new to working with ALTS.

Is there an issue with implementing this inside ALTS? ALTS has access both the underlying socket (so it can set SO_RCVLOWAT) and information (incoming ALTS message length).